hotel-concierge/pyproject.toml
[project]
name = "hotel-concierge"
version = "0.1.0"
description = "AI-powered hotel concierge multi-agent system built on Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=concierge --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "concierge/prompts/*", "concierge/app.py"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
